requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
feedparser==6.0.10
python-dotenv==1.0.0
aiofiles==23.2.1
legacy-cgi==2.6.4
